Cannot access view from service manager when in a script
The module class calls the $serviceManager->get('view') which throws an
exception.  Could put a PHP_SAPI check in the module but we might want
to render templates in a script e.g. html emails. This stops the error
being thrown and normal rendering still works... so... i guess this
is a fix... :D